strides

+++ Source/cuda/linear_forward.cu
@@ -1,10 +1,11 @@
 #include <cuda_runtime.h>
+#include <stdio.h>
 
 __global__ void linear_forward_kernel(
-    const float* input,
-    const float* weights,
-    const float* bias,
-    float* output,
+    const double* input,
+    const double* weights,
+    const double* bias,
+    double* output,
     int batch_size,
     int in_size,
     int out_size)
@@ -13,7 +14,7 @@ __global__ void linear_forward_kernel(
     int x = blockIdx.x * blockDim.x + threadIdx.x; // output neuron
 
     if (i < batch_size && x < out_size) {
-        float sum = bias[x];
+        double sum = bias[x];
         for (int y = 0; y < in_size; ++y) {
             sum += input[i * in_size + y] * weights[x * in_size + y];
         }
@@ -22,10 +23,10 @@ __global__ void linear_forward_kernel(
 }
 
 void launch_linear_forward(
-    const float* d_input,
-    const float* d_weights,
-    const float* d_bias,
-    float* d_output,
+    const double* d_input,
+    const double* d_weights,
+    const double* d_bias,
+    double* d_output,
     int batch_size,
     int in_size,
     int out_size)
